move the SyncMap to util/generic
